Removed more of the old system.
