+++ SS14.Client/GameObjects/Components/Input/ClickableComponent.cs
@@ -4,6 +4,7 @@ using SS14.Shared.GameObjects;
 using SS14.Shared.Interfaces.GameObjects;
 using SS14.Shared.Interfaces.GameObjects.Components;
 using System;
+using SS14.Shared.Enums;
 using SS14.Shared.Map;
 
 namespace SS14.Client.GameObjects
