--- SS14.Client/GameObjects/ContextMenuEntry.cs
@@ -1,9 +1,0 @@
-﻿namespace SS14.Client.GameObjects
-{
-    public struct ContextMenuEntry
-    {
-        public string ComponentMessage { get; set; }
-        public string EntryName { get; set; }
-        public string IconName { get; set; }
-    }
-}
