+++ SS14.Client/Interfaces/GameObjects/Components/IClientClickableComponent.cs
@@ -1,4 +1,5 @@
 ﻿using SS14.Client.State.States;
+using SS14.Shared.Enums;
 using SS14.Shared.GameObjects;
 using SS14.Shared.Interfaces.GameObjects;
 using SS14.Shared.Interfaces.GameObjects.Components;
