+++ SS14.Server/GameObjects/Components/Actor/BasicActorComponent.cs
@@ -8,27 +8,5 @@ namespace SS14.Server.GameObjects
     {
         public override string Name => "BasicActor";
         public IPlayerSession playerSession { get; internal set; }
-
-        public override ComponentReplyMessage ReceiveMessage(object sender, ComponentMessageType type,
-                                                             params object[] list)
-        {
-            ComponentReplyMessage reply = base.ReceiveMessage(sender, type, list);
-
-            if (sender == this)
-                return ComponentReplyMessage.Empty;
-
-            switch (type)
-            {
-                case ComponentMessageType.GetActorConnection:
-                    reply = new ComponentReplyMessage(ComponentMessageType.ReturnActorConnection,
-                                                      playerSession.ConnectedClient);
-                    break;
-                case ComponentMessageType.GetActorSession:
-                    reply = new ComponentReplyMessage(ComponentMessageType.ReturnActorSession, playerSession);
-                    break;
-            }
-
-            return reply;
-        }
     }
 }
