+++ SS14.Shared/Enums/ClickType.cs
@@ -1,0 +1,15 @@
+﻿namespace SS14.Shared.Enums
+{
+    public enum ClickType
+    {
+        None = 0,
+        Left = 1,
+        Right = 2,
+        LeftAlt = 3,
+        RightAlt = 4,
+        LeftShift = 5,
+        RightShift = 6,
+        LeftCtrl = 7,
+        RightCtrl = 8
+    }
+}
