+++ SS14.Shared/GameObjects/EntityEvents.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using SS14.Shared.Enums;
 
 namespace SS14.Shared.GameObjects
 {
